@@ -1,7 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import App from './jsx/App.jsx';
 
-const wrapper = document.getElementById('app-root');
-wrapper ? ReactDOM.render(<App />, wrapper) : false;
-
+import { createRoot } from 'react-dom/client';
+const container = document.getElementById('app-root-2022-ukraine_brief_2');
+const root = createRoot(container);
+root.render(<App/>);
